[Metadata Manager] add function to access metadata
This commit adds Metadata API to access hash objects.

ONE-vscode-DCO-1.0-Signed-off-by: Cookie-Rook <khlee025@gmail.com>
