+++ src/extension.ts
@@ -26,10 +26,10 @@ import {OneTreeDataProvider} from './OneExplorer/OneExplorer';
 import {PartEditorProvider} from './PartEditor/PartEditor';
 import {PartGraphSelPanel} from './PartEditor/PartGraphSelector';
 import { ToolchainProvider } from './Toolchain/ToolchainProvider';
-import { Metadata } from './Metaprovider/metadata_manager';
+import { Metadata } from './MetadataManager/metadataAPI';
 import {Logger} from './Utils/Logger';
 
-import { MetadataEventManager } from './FileManager/EventManager';
+import { MetadataEventManager } from './MetadataManager/EventManager';
 
 /* istanbul ignore next */
 export function activate(context: vscode.ExtensionContext) {
